+++ echoframework/src/main/kotlin/org/echo/mobile/framework/facade/FeeFacade.kt
@@ -2,6 +2,7 @@ package org.echo.mobile.framework.facade
 
 import org.echo.mobile.framework.Callback
 import org.echo.mobile.framework.model.AssetAmount
+import org.echo.mobile.framework.model.TransactionResult
 import org.echo.mobile.framework.model.contract.ContractFee
 import org.echo.mobile.framework.model.contract.input.InputValue
 
@@ -94,6 +95,40 @@ interface FeeFacade {
         assetId: String,
         feeAsset: String?,
         callback: Callback<AssetAmount>
+    )
+
+    /**
+     * Counts required fee for ethereum erc20 withdraw operation
+     */
+    fun getFeeForWithdrawErc20Operation(
+            accountNameOrId: String,
+            ethAddress: String,
+            ethTokenId: String,
+            value: String,
+            feeAsset: String,
+            callback: Callback<AssetAmount>
+    )
+
+    /**
+     * Counts required fee for ethereum withdraw operation
+     */
+    fun getFeeForWithdrawEthereumOperation(
+            accountNameOrId: String,
+            ethAddress: String,
+            value: String,
+            feeAsset: String,
+            callback: Callback<AssetAmount>
+    )
+
+    /**
+     * Counts required fee for btc withdraw transaction
+     */
+    fun getFeeForWithdrawBtcOperation(
+            accountNameOrId: String,
+            btcAddress: String,
+            value: String,
+            feeAsset: String,
+            callback: Callback<AssetAmount>
     )
 
 }